backend/requirements.txt
# 后端最小运行依赖会在 XIV-004 阶段正式落定
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic-settings==2.5.2
httpx==0.27.2
pypdf==5.0.1
pytest==8.3.3
